Support batching oplog replay requests
Before this change when replaying the oplog we maxed out at ~150
requests per second. We think the bottleneck was the round trip time
between the database and the oplog-replay script. To address that issue
we're supporting batching oplog requests.

Conceptually we still have one goroutine responsible for putting the ops
in the "queue" to be processed, but now the goroutine that actual calls
mongo to apply to operation gets all the ops it has in its queue and
applies them at once.